Moved password reset logic in its own package
