+++ iam-login-service/src/main/java/it/infn/mw/iam/api/account/password_reset/DefaultPasswordResetService.java
@@ -1,4 +1,4 @@
-package it.infn.mw.iam.api.account;
+package it.infn.mw.iam.api.account.password_reset;
 
 import java.util.NoSuchElementException;
