+++ iam-login-service/src/main/java/it/infn/mw/iam/api/account/password_reset/EmailDTO.java
@@ -1,4 +1,4 @@
-package it.infn.mw.iam.api.account;
+package it.infn.mw.iam.api.account.password_reset;
 
 import javax.validation.constraints.NotNull;
